# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Capture API Simulator API",
    "description": "Deterministic traffic-forensics simulator. Browsers cannot call it directly (no CORS): proxy it from your own server. Optional fields are absent rather than null; times are ISO-8601 UTC with milliseconds; session ids are uint64 decimal strings.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/meta/fields": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Filter fields the search accepts",
        "description": "The filter catalogue: every field, its type, the operators it allows and, where it is closed, its enum. Build filter UI from this rather than from a copy in the client. The `sensor` field's enum includes ready imports.",
        "operationId": "listFields",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldList"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, invalid or revoked credentials",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge (Bearer-authenticated routes).",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/meta/columns": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Grid columns in their default order",
        "description": "The session grid's columns, in default order, with default visibility, sortability and a width hint. One column carries a `type` that is not in the documented list: render unknown types as text.",
        "operationId": "listColumns",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ColumnList"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, invalid or revoked credentials",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge (Bearer-authenticated routes).",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/meta/enums/{name}": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Values of one closed enum",
        "description": "Values and labels of a closed field. The FIRST `country` call of each token family answers 503 `catalog_warming` with `Retry-After: 2` while the geo catalogue loads — retry it once and it succeeds.",
        "operationId": "getEnum",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "One of protocol, country, risk_band, detection_rule, dns_query_type, dns_rcode, http_method, tls_version, smb2_status",
              "title": "Name"
            },
            "description": "One of protocol, country, risk_band, detection_rule, dns_query_type, dns_rcode, http_method, tls_version, smb2_status"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnumResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, invalid or revoked credentials",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge (Bearer-authenticated routes).",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Temporarily unavailable; honour Retry-After",
            "headers": {
              "Retry-After": {
                "description": "Integer seconds (an HTTP-date on /v1/auth/login).",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/meta/schema/{protocol}": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Declared fields of a decoded protocol",
        "description": "The paths a decoded payload declares for this protocol, for both decoder generations. The decoder emits a few fields that are NOT declared here (for example `tls.ja4`): show them, but show them as undeclared.",
        "operationId": "getProtocolSchema",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "name": "protocol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Protocol name, e.g. `dns`.",
              "title": "Protocol"
            },
            "description": "Protocol name, e.g. `dns`."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolSchema"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, invalid or revoked credentials",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge (Bearer-authenticated routes).",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ColumnDef": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "type": {
            "type": "string",
            "enum": [
              "ts",
              "ip_port",
              "bytes",
              "risk",
              "protocol",
              "duration",
              "text",
              "country",
              "sensor",
              "id"
            ],
            "title": "Type",
            "description": "Documented column types. Render any other value as text."
          },
          "default_visible": {
            "type": "boolean",
            "title": "Default Visible"
          },
          "sortable": {
            "type": "boolean",
            "title": "Sortable"
          },
          "width_hint": {
            "type": "integer",
            "title": "Width Hint"
          }
        },
        "type": "object",
        "required": [
          "key",
          "label",
          "type",
          "default_visible",
          "sortable",
          "width_hint"
        ],
        "title": "ColumnDef"
      },
      "ColumnList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ColumnDef"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ColumnList"
      },
...
      "DecoderVersion": {
        "type": "string",
        "enum": [
          "v1",
          "v2"
        ]
      },
...
      "EnumName": {
        "type": "string",
        "enum": [
          "protocol",
          "country",
          "risk_band",
          "detection_rule",
          "dns_query_type",
          "dns_rcode",
          "http_method",
          "tls_version",
          "smb2_status"
        ]
      },
      "EnumResponse": {
        "properties": {
          "name": {
            "$ref": "#/components/schemas/EnumName"
          },
          "values": {
            "items": {
              "$ref": "#/components/schemas/EnumValue"
            },
            "type": "array",
            "title": "Values"
          }
        },
        "type": "object",
        "required": [
          "name",
          "values"
        ],
        "title": "EnumResponse"
      },
      "EnumValue": {
        "properties": {
          "value": {
            "type": "string",
            "title": "Value"
          },
          "label": {
            "type": "string",
            "title": "Label"
          }
        },
        "type": "object",
        "required": [
          "value",
          "label"
        ],
        "title": "EnumValue"
      },
      "ErrorBody": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Human-readable message."
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable snake_case error code."
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorBody",
        "description": "Error envelope of every domain error. Extra keys carry machine-readable context."
      },
...
      "FieldDef": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "type": {
            "$ref": "#/components/schemas/FieldType"
          },
          "operators": {
            "items": {
              "$ref": "#/components/schemas/FilterOp"
            },
            "type": "array",
            "title": "Operators"
          },
          "enum": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Enum"
          },
          "enum_name": {
            "type": "string",
            "title": "Enum Name",
            "description": "The `/v1/meta/enums/{name}` catalogue that carries display labels for this field's values. Absent when the field has no catalogue entry (`sensor`), so the route name is never guessed from the field name."
          },
          "pattern": {
            "type": "string",
            "title": "Pattern"
          },
          "example": {
            "type": "string",
            "title": "Example"
          }
        },
        "type": "object",
        "required": [
          "name",
          "label",
          "type",
          "operators",
          "example"
        ],
        "title": "FieldDef"
      },
      "FieldList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FieldDef"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "FieldList"
      },
      "FieldType": {
        "type": "string",
        "enum": [
          "ip",
          "cidr",
          "port",
          "number",
          "bytes",
          "string",
          "enum",
          "country",
          "ja3",
          "duration_ms",
          "sensor"
        ]
      },
...
      "FilterOp": {
        "type": "string",
        "enum": [
          "eq",
          "in",
          "cidr",
          "glob",
          "gte",
          "lte",
          "between",
          "exists"
        ]
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProtocolName": {
        "type": "string",
        "enum": [
          "dns",
          "http",
          "tls",
          "smtp",
          "smb2",
          "ssh",
          "ntp",
          "tcp"
        ]
      },
      "ProtocolSchema": {
        "properties": {
          "protocol": {
            "$ref": "#/components/schemas/ProtocolName"
          },
          "decoder_versions": {
            "items": {
              "$ref": "#/components/schemas/DecoderVersion"
            },
            "type": "array",
            "title": "Decoder Versions"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/SchemaField"
            },
            "type": "array",
            "title": "Fields"
          }
        },
        "type": "object",
        "required": [
          "protocol",
          "decoder_versions",
          "fields"
        ],
        "title": "ProtocolSchema"
      },
...
      "SchemaField": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path",
            "description": "Dotted path inside `decoded`, e.g. \"dns.query.name\"."
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "sensitive": {
            "type": "boolean",
            "title": "Sensitive"
          }
        },
        "type": "object",
        "required": [
          "path",
          "title",
          "type"
        ],
        "title": "SchemaField"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "description": "Access token `at_…` from POST /v1/auth/login or /v1/auth/refresh.",
        "scheme": "bearer"
      }
    }
  },
...
}